mostly done (animation not working)
